requirements.txt
streamlit
streamlit-option-menu
scikit-learn
mysql-connector-python
